work done with test

+++ src/main/java/com/project/TopTierRentals/models/Customer.java
@@ -12,7 +12,7 @@ public class Customer {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int customerId;
+    private Long customerId;
     private String customerName;
     private String customerAddress;
     private String customerPhone;
